+++ CSI-340_ChenRoblesWu_Website/CSI-340_ChenRoblesWu_Website/Controllers/CartController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using CSI_340_ChenRoblesWu_Website.Data;
+using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,9 +9,32 @@ namespace CSI_340_ChenRoblesWu_Website.Controllers
 {
     public class CartController : Controller
     {
-        public IActionResult ShoppingCart()
+        private readonly DataContext _db;
+
+        public CartController(DataContext db)
+        {
+            _db = db;
+        }
+        public IActionResult Info()
+        {
+            var displayData = _db.Book.ToList();
+            return View( "ShoppingCart",displayData);
+        }
+        public IActionResult Create()
         {
             return View();
+        }
+        [HttpPost]
+
+        public async Task<IActionResult> Create(int book_id)
+        {
+            if (ModelState.IsValid)
+            {
+                _db.Add(book_id);
+                await _db.SaveChangesAsync();
+                return RedirectToAction("ContactInfo");
+            }
+            return View(book_id);
         }
     }
 }
