+++ CSI-340_ChenRoblesWu_Website/CSI-340_ChenRoblesWu_Website/Views/Account/Login.cshtml
@@ -11,12 +11,7 @@
 <div class="row">
     <div class="col-md-4">
         <form asp-action="Login">
-            <!--<div asp-validation-summary="ModelOnly" class="text-danger"></div>
-            <div class="form-group">
-            <label asp-for="StudentId" class="control-label"></label>
-            <input asp-for="StudentId" class="form-control" />
-            <span asp-validation-for="StudentId" class="text-danger"></span>
-            </div>-->
+            
             <div class="form-group">
                 <label asp-for="FirstName" class="control-label"></label>
                 <input asp-for="FirstName" class="form-control" />
