+++ CSI-340_ChenRoblesWu_Website/CSI-340_ChenRoblesWu_Website/Views/Shared/_Layout.cshtml
@@ -49,7 +49,7 @@
                                 <a class="nav-link text-dark" asp-area="" asp-controller="Account" asp-action="Login">Login</a>
                             </li>
                             <li class="nav-item">
-                                <a class="nav-link text-dark" asp-area="" asp-controller="Account" asp-action="info">Cart</a>
+                                <a class="nav-link text-dark" asp-area="" asp-controller="Cart" asp-action="info">Cart</a>
                             </li>
                     </ul>
                 </div>
